pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["TAYLOR JOHNSON <TAYLOR.MCMINN@GMAIL.COM>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
